requirements.txt
pipdeptree
pip-licenses
python-dotenv
ipykernel

pycodestyle
autopep8

openai
tiktoken
pandas
